requirements.txt
anyio==4.12.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
exceptiongroup==1.3.1
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.4
googleapis-common-protos==1.72.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
oauthlib==3.3.1
proto-plus==1.27.0
protobuf==6.33.4
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pyparsing==3.3.2
python-dotenv==1.2.1
python-telegram-bot==22.6
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
